pyproject.toml
[tool.poetry]
name = "portfolio_site"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
Pillow = "^9.1.1"
django-ckeditor = "^6.4.2"
PyJWT = "^2.4.0"
djangorestframework-simplejwt = "^5.2.0"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.20.0"
django-ratelimit = "^3.0.1"
redis = "^4.3.4"
psycopg2 = "^2.9.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
